refactor: ディレクトリ構成見直し
